feat: 全面迁移优化 — meter/thermal/system 模块实体与服务完善、组织权限拦截器、清理旧SQL脚本
- meter 模块：所有仪表档案 Entity/Vo/Bo/Mapper/Service 完善字段映射与查询
- thermal 模块：物业收费/热力调控/微信支付/IoT/定时任务 全链路优化
- system 模块：租户管理/菜单/区域/日志服务完善
- common 模块：新增 OrgPermission 组织权限注解与拦截器
- 新增 tenant_db_schema.sql 独立维护
- 清理旧版 RuoYi 多数据库 SQL 脚本和审计报告

+++ sdkj-modules/sdkj-thermal/src/main/java/org/sdkj/thermal/quartz/ThermalJobManager.java
@@ -19,7 +19,7 @@ public class ThermalJobManager {
     private static final String JOB_GROUP = "THERMAL_JOBS";
     private static final String TRIGGER_GROUP = "THERMAL_TRIGGERS";
 
-    public boolean addJob(Integer taskId) throws SchedulerException {
+    public boolean addJob(Long taskId) throws SchedulerException {
         HtTasks task = tasksMapper.selectById(taskId);
         if (task == null) {
             log.warn("任务不存在: {}", taskId);
@@ -65,7 +65,7 @@ public class ThermalJobManager {
         return true;
     }
 
-    public boolean resumeJob(Integer taskId) throws SchedulerException {
+    public boolean resumeJob(Long taskId) throws SchedulerException {
         JobKey jobKey = jobKey(TenantContextHolder.getTenantCode(), taskId);
         if (!scheduler.checkExists(jobKey)) {
             return addJob(taskId);
@@ -75,7 +75,7 @@ public class ThermalJobManager {
         return true;
     }
 
-    public boolean pauseJob(Integer taskId) throws SchedulerException {
+    public boolean pauseJob(Long taskId) throws SchedulerException {
         JobKey jobKey = jobKey(TenantContextHolder.getTenantCode(), taskId);
         if (scheduler.checkExists(jobKey)) {
             scheduler.pauseJob(jobKey);
@@ -84,7 +84,7 @@ public class ThermalJobManager {
         return true;
     }
 
-    public boolean deleteJob(Integer taskId) throws SchedulerException {
+    public boolean deleteJob(Long taskId) throws SchedulerException {
         JobKey jobKey = jobKey(TenantContextHolder.getTenantCode(), taskId);
         if (scheduler.checkExists(jobKey)) {
             scheduler.deleteJob(jobKey);
@@ -93,7 +93,7 @@ public class ThermalJobManager {
         return true;
     }
 
-    public boolean triggerJob(Integer taskId) throws SchedulerException {
+    public boolean triggerJob(Long taskId) throws SchedulerException {
         JobKey jobKey = jobKey(TenantContextHolder.getTenantCode(), taskId);
         if (scheduler.checkExists(jobKey)) {
             scheduler.triggerJob(jobKey);
@@ -103,11 +103,11 @@ public class ThermalJobManager {
         return addJob(taskId);
     }
 
-    private JobKey jobKey(String tenantCode, Integer taskId) {
+    private JobKey jobKey(String tenantCode, Long taskId) {
         return new JobKey("thermal_" + tenantKey(tenantCode) + "_" + taskId, JOB_GROUP);
     }
 
-    private TriggerKey triggerKey(String tenantCode, Integer taskId) {
+    private TriggerKey triggerKey(String tenantCode, Long taskId) {
         return new TriggerKey("trigger_" + tenantKey(tenantCode) + "_" + taskId, TRIGGER_GROUP);
     }
